[Move] new path for `section.ts`
